move commit parsing logic to gitrepo package
