+++ internal/librarian/release_please_lite_test.go
@@ -27,149 +27,6 @@ import (
 	"github.com/googleapis/librarian/internal/gitrepo"
 )
 
-func TestParseCommit(t *testing.T) {
-	for _, test := range []struct {
-		name    string
-		message string
-		want    *ConventionalCommit
-		wantErr bool
-	}{
-		{
-			name:    "simple feat",
-			message: "feat: add new feature",
-			want: &ConventionalCommit{
-				Type:        "feat",
-				Description: "add new feature",
-				Footers:     make(map[string]string),
-				SHA:         "fake-sha",
-			},
-		},
-		{
-			name:    "feat with scope",
-			message: "feat(scope): add new feature",
-			want: &ConventionalCommit{
-				Type:        "feat",
-				Scope:       "scope",
-				Description: "add new feature",
-				Footers:     make(map[string]string),
-				SHA:         "fake-sha",
-			},
-		},
-		{
-			name:    "feat with breaking change",
-			message: "feat!: add new feature",
-			want: &ConventionalCommit{
-				Type:        "feat",
-				Description: "add new feature",
-				IsBreaking:  true,
-				Footers:     make(map[string]string),
-				SHA:         "fake-sha",
-			},
-		},
-		{
-			name:    "feat with single footer",
-			message: "feat: add new feature\n\nCo-authored-by: John Doe <john.doe@example.com>",
-			want: &ConventionalCommit{
-				Type:        "feat",
-				Description: "add new feature",
-				Footers:     map[string]string{"Co-authored-by": "John Doe <john.doe@example.com>"},
-				SHA:         "fake-sha",
-			},
-		},
-		{
-			name:    "feat with multiple footers",
-			message: "feat: add new feature\n\nCo-authored-by: John Doe <john.doe@example.com>\nReviewed-by: Jane Smith <jane.smith@example.com>",
-			want: &ConventionalCommit{
-				Type:        "feat",
-				Description: "add new feature",
-				Footers: map[string]string{
-					"Co-authored-by": "John Doe <john.doe@example.com>",
-					"Reviewed-by":    "Jane Smith <jane.smith@example.com>",
-				},
-				SHA: "fake-sha",
-			},
-		},
-		{
-			name: "feat with multiple footers for generated changes",
-			message: `feat: [library-name] add new feature
-This is the body.
-...
-
-PiperOrigin-RevId: piper_cl_number
-
-Source-Link: [googleapis/googleapis@{source_commit_hash}](https://github.com/googleapis/googleapis/commit/{source_commit_hash})
-`,
-			want: &ConventionalCommit{
-				Type:        "feat",
-				Description: "[library-name] add new feature",
-				Body:        "This is the body.\n...",
-				IsBreaking:  false,
-				Footers: map[string]string{
-					"PiperOrigin-RevId": "piper_cl_number",
-					"Source-Link":       "[googleapis/googleapis@{source_commit_hash}](https://github.com/googleapis/googleapis/commit/{source_commit_hash})",
-				},
-				SHA: "fake-sha",
-			},
-		},
-		{
-			name:    "feat with breaking change footer",
-			message: "feat: add new feature\n\nBREAKING CHANGE: this is a breaking change",
-			want: &ConventionalCommit{
-				Type:        "feat",
-				Description: "add new feature",
-				Body:        "",
-				IsBreaking:  true,
-				Footers:     map[string]string{"BREAKING CHANGE": "this is a breaking change"},
-				SHA:         "fake-sha",
-			},
-		},
-		{
-			name:    "feat with body and footers",
-			message: "feat: add new feature\n\nThis is the body of the commit message.\nIt can span multiple lines.\n\nCo-authored-by: John Doe <john.doe@example.com>",
-			want: &ConventionalCommit{
-				Type:        "feat",
-				Description: "add new feature",
-				Body:        "This is the body of the commit message.\nIt can span multiple lines.",
-				Footers:     map[string]string{"Co-authored-by": "John Doe <john.doe@example.com>"},
-				SHA:         "fake-sha",
-			},
-		},
-		{
-			name:    "feat with multi-line footer",
-			message: "feat: add new feature\n\nThis is the body.\n\nBREAKING CHANGE: this is a breaking change\nthat spans multiple lines.",
-			want: &ConventionalCommit{
-				Type:        "feat",
-				Description: "add new feature",
-				Body:        "This is the body.",
-				IsBreaking:  true,
-				Footers:     map[string]string{"BREAKING CHANGE": "this is a breaking change\nthat spans multiple lines."},
-				SHA:         "fake-sha",
-			},
-		},
-		{
-			name:    "invalid commit",
-			message: "this is not a conventional commit",
-			wantErr: true,
-		},
-	} {
-		t.Run(test.name, func(t *testing.T) {
-			got, err := parseCommit(test.message, "fake-sha")
-			if test.wantErr {
-				if err == nil {
-					t.Errorf("%s should return error", test.name)
-				}
-				return
-			}
-			if err != nil {
-				t.Fatal(err)
-			}
-			if diff := cmp.Diff(test.want, got); diff != "" {
-				t.Errorf("ParseCommit(%q) returned diff (-want +got):\n%s", test.message, diff)
-			}
-		})
-	}
-}
-
 func TestShouldExclude(t *testing.T) {
 	testCases := []struct {
 		name         string
@@ -311,7 +168,7 @@ func TestGetConventionalCommitsSinceLastRelease(t *testing.T) {
 	for _, test := range []struct {
 		name    string
 		library *config.LibraryState
-		want    []*ConventionalCommit
+		want    []*gitrepo.ConventionalCommit
 		wantErr bool
 	}{
 		{
@@ -323,7 +180,7 @@ func TestGetConventionalCommitsSinceLastRelease(t *testing.T) {
 				SourceRoots:         []string{"foo"},
 				ReleaseExcludePaths: []string{"foo/README.md"},
 			},
-			want: []*ConventionalCommit{
+			want: []*gitrepo.ConventionalCommit{
 				{
 					Type:        "feat",
 					Scope:       "foo",
@@ -345,7 +202,7 @@ func TestGetConventionalCommitsSinceLastRelease(t *testing.T) {
 				t.Errorf("GetCommits() error = %v, wantErr %v", err, test.wantErr)
 				return
 			}
-			if diff := cmp.Diff(test.want, got, cmpopts.IgnoreFields(ConventionalCommit{}, "SHA", "Body", "IsBreaking")); diff != "" {
+			if diff := cmp.Diff(test.want, got, cmpopts.IgnoreFields(gitrepo.ConventionalCommit{}, "SHA", "Body", "IsBreaking")); diff != "" {
 				t.Errorf("GetCommits() mismatch (-want +got):\n%s", diff)
 			}
 		})
